add subscription_log and fix unsub/pause

+++ app/controllers/subscriptions_controller.rb
@@ -2,11 +2,12 @@ class SubscriptionsController < ApplicationController
   # 購読
   def create
     service = GoogleDirectoryService.instance
-    member = Google::Apis::AdminDirectoryV1::Member.new(email: params[:email])
 
     begin
+      member = Google::Apis::AdminDirectoryV1::Member.new(email: params[:email])
       service.insert_member('test@notsobad.jp', member)
       flash[:success] = '登録完了しました！明日の配信からメールが届きます。<a href="https://blog.bungomail.com/entry/2018/05/21/172542" target="_blank" class="text-link">メールの受信設定</a>を事前に必ずご確認ください。'
+      SubscriptionLog.create!(email: params[:email], action: "subscribed")
     rescue => e
       flash[:error] = case e.status_code
                       when 409
@@ -20,40 +21,23 @@ class SubscriptionsController < ApplicationController
     redirect_to root_path
   end
 
-  # 月末まで一時停止
-  def update
-    month = Time.current.strftime("%Y%m")
-    paused_log = PausedLog.find_or_initialize_by(email: params[:email], month: month)
-    if paused_log.persisted?
-      redirect_to page_path(:unsubscribe), flash: { error: "このアドレスはすでに月末まで配信停止中です。もし配信が止まらない場合はinfo@notsobad.jpまでお問い合わせください。" } and return
-    end
-
-    service = GoogleDirectoryService.instance
-    member = Google::Apis::AdminDirectoryV1::Member.new(delivery_settings: 'NONE')
-
-    begin
-      service.patch_member('test@notsobad.jp', params[:email], member)
-      paused_log.save!
-      flash[:success] = '月末まで配信を一時停止しました。来月になると自動で配信が再開されます。'
-      redirect_to root_path
-    rescue => e
-      flash[:error] = case e.status_code
-                      when 404
-                        'メールアドレスが見つかりませんでした。。再度試してもうまくいかないない場合、お手数ですがinfo@notsobad.jpまでお問い合わせください。'
-                      else
-                        e.message
-                      end
-      redirect_to page_path(:unsubscribe)
-    end
-  end
-
-  # 登録解除
+  # 配信停止
   def destroy
     service = GoogleDirectoryService.instance
 
     begin
-      service.delete_member('test@notsobad.jp', params[:email])
-      flash[:success] = '購読を解除しました。明日の配信からメールが届かなくなります。これまでのご利用ありがとうございました。'
+      # 登録解除
+      if params[:action_type] == 'unsubscribed'
+        service.delete_member('test@notsobad.jp', params[:email])
+        flash[:success] = '購読を解除しました。明日の配信からメールが届かなくなります。これまでのご利用ありがとうございました。'
+      # 月末まで一時停止
+      elsif params[:action_type] == 'paused'
+        member = Google::Apis::AdminDirectoryV1::Member.new(delivery_settings: 'NONE')
+        service.patch_member('test@notsobad.jp', params[:email], member)
+        flash[:success] = '月末まで配信を一時停止しました。来月になると自動で配信が再開されます。'
+      end
+
+      SubscriptionLog.create!(email: params[:email], action: params[:action_type])
       redirect_to root_path
     rescue => e
       flash[:error] = case e.status_code
